fix(formula): correct build install logic for chezroot
The `std_go_args` helper handles both building and installing the
binary, so the redundant `bin.install` line was causing the build
to fail. This commit removes it.

ci: remove workflow as tap is now updated by goreleaser

Removes the 'release.yml' workflow, which was triggered by a
workflow_dispatch. The release process in the 'main-branch/chezroot'
repository has been updated to use goreleaser, which now pushes
updates to this formula directly.